refactor(generator): split translator vs. parser
This change introduces a new interface `Parser`. Its job is to convert the
textual representation of an API specification (i.e., the OpenAPI or Protobuf
inputs) into the `genclient.API` AST.

In this PR we create the parser and split the Protobuf and OpenAPI translators
to depend on them. A future PR will create a single standalone program to call
the parser and per-language Codec.
