+++ generator/internal/genclient/genclient.go
@@ -95,6 +95,30 @@ type LanguageCodec interface {
 	FormatDocComments(string) []string
 }
 
+// Parser converts an textual specification to a `genclient.API` object.
+//
+// The generator supports both Protobuf and OpenAPI v3 specifications. The
+// parser layer converts a textual specification of the model into the
+// an `genclient.API` object. Basically an AST of the API. Other layers convert
+// this into generated code.
+type Parser interface {
+	// A help message describing the parser.
+	Help() string
+	// Returns a map of known options and their description.
+	OptionDescriptions() map[string]string
+	// Parses the API specification and returns the API as a syntax tree.
+	Parse(config ParserOptions) (*API, error)
+}
+
+type ParserOptions struct {
+	// The location where the specification can be found.
+	Source string
+	// The location of the service configuration file.
+	ServiceConfig string
+	// Additional options.
+	Options map[string]string
+}
+
 // GenerateRequest used to generate clients.
 type GenerateRequest struct {
 	// The in memory representation of a parsed input.
