+++ generator/internal/genclient/parser/protobuf/annotations.go
