+++ generator/internal/genclient/parser/protobuf/testdata/enum.proto
