+++ generator/internal/genclient/parser/protobuf/testdata/object_fields.proto
