+++ generator/internal/genclient/parser/protobuf/testdata/oneofs.proto
