+++ generator/internal/genclient/parser/protobuf/testdata/scalar.proto
