+++ generator/internal/genclient/parser/protobuf/testdata/scalar_array.proto
