+++ generator/internal/genclient/parser/protobuf/testdata/scalar_optional.proto
