+++ generator/internal/genclient/parser/protobuf/testdata/with_import.proto
