+++ generator/internal/genclient/parser/protobuf/testdata/wkt_fields.proto
